add services section and some desing
Add services section and some design

This commit adds a services section to the website and incorporates some
design changes. The services section includes three cards with icons and
information about front-end web development, image design, and web
management. The design changes include styling adjustments to the cards
and the addition of a title for the services section.